Refactors to avoid two conns during seeding
