+++ database/auth.js
@@ -1,0 +1,8 @@
+const auth = {
+  database: 'fRiend',
+  user: 'test',
+  password: 'test',
+  host: 'localhost',
+};
+
+module.exports = auth;
